Lab9

+++ lab9/Core/ExhibitionFacade.cs
@@ -9,15 +9,15 @@ public class ExhibitionFacade
 {
     private readonly AppDbContext _context;
     private readonly ExhibitionRepository _exhibitionRepository;
-    private readonly VisitorRepository _visitorRepository;
+    private readonly NamesGetter _namesGetter;
     private readonly TicketRepository _ticketRepository;
 
     public ExhibitionFacade(AppDbContext context)
     {
         _context = context;
         _exhibitionRepository = new ExhibitionRepository(context);
-        _visitorRepository = new VisitorRepository(context);
         _ticketRepository = new TicketRepository(context);
+        _namesGetter = new NamesGetter(context);
     }
     
     public async Task AddExhibition()
@@ -25,10 +25,15 @@ public class ExhibitionFacade
         try
         {
             Console.Write("Введите название выставки: ");
-            var name = Console.ReadLine() ?? "Concert";
-            Console.Write("Enter Start Date (yyyy-MM-dd): ");
-            var date = DateTime.Parse(Console.ReadLine() ?? "01.01.2024 00:00:00").ToUniversalTime();
-                        
+            var name = Console.ReadLine()!;
+            if (name.Length < 3)
+            {
+                throw new Exception("Название не может быть короче 3 символов");
+            }
+            
+            Console.Write("Введите дату проведения выставки (дд-мм-гггг): ");
+            var date = DateTime.Parse(Console.ReadLine()!).ToUniversalTime() ;
+            
             var exhibition = Exhibition.Create(name, date);
             Console.WriteLine($"Создана выставка {exhibition.Id}");
             
@@ -37,11 +42,10 @@ public class ExhibitionFacade
         catch (FormatException)
         {
             Console.WriteLine("Неверный формат вводимых данных");
-            throw;
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
+            Console.WriteLine(e.Message);
         }
     }
 
@@ -49,19 +53,31 @@ public class ExhibitionFacade
     {
         try
         {
+            await GetAllExhibitions();
+            
             Console.Write("Введите ID выставки: ");
-            var id = Guid.Parse(Console.ReadLine());
-            await _exhibitionRepository.Delete(id);
-            Console.WriteLine("Выставка удалена");
+            var id = Guid.Parse(Console.ReadLine()!);
+            
+            var guids = await _namesGetter.GetExhibitionGuidsList();
+            if (guids.Any(g => g == id))
+            {
+                await _ticketRepository.DeleteByExhibitionId(id);
+                Console.WriteLine("Билеты на выставку удалены");
+                await _exhibitionRepository.Delete(id);
+                Console.WriteLine("Выставка удалена");
+            }
+            else
+            {
+                Console.WriteLine("Выставка не найдена");
+            }
         }
         catch (FormatException)
         {
-            Console.WriteLine("Неверный формат вводимых данных");
-            throw;
+            Console.WriteLine("Неверный формат ID");
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
+            Console.WriteLine(e.Message);
         }
     }
     
@@ -69,23 +85,31 @@ public class ExhibitionFacade
     {
         try
         {
+            await GetAllExhibitions();
+            
             Console.Write("Введите ID выставки: ");
-            var id = Guid.Parse(Console.ReadLine());
+            var id = Guid.Parse(Console.ReadLine()!);
+            
             Console.Write("Введите название выставки: ");
-            var name = Console.ReadLine() ?? "Concert";
+            var name = Console.ReadLine()!;
+            if (name.Length < 2)
+            {
+                throw new Exception("Название не может быть короче 3 символов");
+            }
+            
             Console.Write("Enter Start Date (yyyy-MM-dd): ");
-            var date = DateTime.Parse(Console.ReadLine() ?? "01.01.2024 00:00:00").ToUniversalTime();
+            var date = DateTime.Parse(Console.ReadLine()!).ToUniversalTime();
+            
             await _exhibitionRepository.Update(id, name, date);
             Console.WriteLine("Выставка обновлена");
         }
         catch (FormatException)
         {
             Console.WriteLine("Неверный формат вводимых данных");
-            throw;
         }
         catch (Exception e)
         {
-            Console.WriteLine(e);
+            Console.WriteLine(e.Message);
         }
     }
     
@@ -113,110 +137,119 @@ public class ExhibitionFacade
     // Сколько билетов продано на выставку
     public async Task GetTicketsSold()
     {
-        await GetAllExhibitionNames();
-     
+        await _namesGetter.GetAllExhibitionNames();
+
         Console.Write("Введите название выставки: ");
         var exhibitionName = Console.ReadLine()!;
         
-        var ticketsQuery =
-            from ticket in _context.Tickets
-            join exhibition in _context.Exhibitions on ticket.ExhibitionId equals exhibition.Id
-            where exhibition.Name == exhibitionName
-            select ticket;
-        
-        var amount = await ticketsQuery.CountAsync();
+        var allExhibitionNames = await _namesGetter.GetExhibitionNamesList();
 
-        Console.WriteLine($"Продано билетов: {amount}\n");
+        if (allExhibitionNames.Any(name => name == exhibitionName))
+        {
+            var ticketsQuery =
+                from ticket in _context.Tickets
+                join exhibition in _context.Exhibitions on ticket.ExhibitionId equals exhibition.Id
+                where exhibition.Name == exhibitionName
+                select ticket;
+
+            var amount = await ticketsQuery.CountAsync();
+
+            Console.WriteLine(
+                amount == 0
+                    ? $"На данную выставку билеты еще не были проданы\n"
+                    : $"Продано билетов: {amount}\n"
+            );
+        }
+        else
+        {
+            Console.WriteLine($"Выставки {exhibitionName} не существует");
+        }
     }
 
     // На сколько уникальных выставок сходил посетитель
     public async Task GetUniqueExhibitionsVisited()
     {
-        await GetAllVisitorsNames();
+        await _namesGetter.GetAllVisitorsNames();
         
         Console.Write("Введите имя посетителя: ");
         var visitorName = Console.ReadLine()!;
-        
-        var exhibitionsQuery =
-            from ticket in _context.Tickets
-            join visitor in _context.Visitors on ticket.VisitorId equals visitor.Id
-            where visitor.FullName == visitorName
-            select ticket.ExhibitionId;
-        
-        var totalSpendQuery =
-            from ticket in _context.Tickets
-            join visitor in _context.Visitors on ticket.VisitorId equals visitor.Id
-            where visitor.FullName == visitorName
-            select ticket.Price * (100 - visitor.Discount) / 100;
-        
-        var distinctAmount = await exhibitionsQuery
-            .Distinct()
-            .CountAsync();
-        
-        var totalAmount = await exhibitionsQuery
-            .CountAsync();
 
-        var totalSpend = await totalSpendQuery
-            .SumAsync();
         
-        Console.WriteLine($"Посещено уникальных выставок: {distinctAmount}");
-        Console.WriteLine($"Куплено всего билетов: {totalAmount}");
-        Console.WriteLine($"Всего потрачено денег на выставки: {totalSpend:0.00}\n");
+        var allVisitorsNames = await _namesGetter.GetVisitorsNamesList();
         
+        if (allVisitorsNames.Any(name => name == visitorName))
+        {
+            var exhibitionsQuery =
+                from ticket in _context.Tickets
+                join visitor in _context.Visitors on ticket.VisitorId equals visitor.Id
+                where visitor.FullName == visitorName
+                select ticket.ExhibitionId;
+        
+            var totalSpendQuery =
+                from ticket in _context.Tickets
+                join visitor in _context.Visitors on ticket.VisitorId equals visitor.Id
+                where visitor.FullName == visitorName
+                select ticket.Price * (100 - visitor.Discount) / 100;
+        
+            var distinctAmount = await exhibitionsQuery
+                .Distinct()
+                .CountAsync();
+        
+            var totalAmount = await exhibitionsQuery
+                .CountAsync();
+
+            var totalSpend = await totalSpendQuery
+                .SumAsync();
+        
+            Console.WriteLine($"Посещено уникальных выставок: {distinctAmount}");
+            Console.WriteLine($"Куплено всего билетов: {totalAmount}");
+            Console.WriteLine($"Всего потрачено денег на выставки: {totalSpend:0.00}\n");
+        }
+        else
+        {
+            Console.WriteLine($"Посетитель {visitorName} не найден");
+        }
+       
     }
 
     // Средний процент скидки на выставку
     public async Task GetAverageDiscount()
     {
-        await GetAllExhibitionNames();
+        await _namesGetter.GetAllExhibitionNames();
         
         Console.Write("Введите название выставки: ");
         var exhibitionName = Console.ReadLine()!;
         
-        var discountsQuery =
-            from ticket in _context.Tickets
-            join visitor in _context.Visitors on ticket.VisitorId equals visitor.Id
-            join exhibition in _context.Exhibitions on ticket.ExhibitionId equals exhibition.Id
-            where exhibition.Name == exhibitionName
-            select visitor.Discount;
-        
-        var discounts = await discountsQuery
-            .Distinct().
-            ToListAsync();
+        var allExhibitionNames = await _namesGetter.GetExhibitionNamesList();
 
-        var avgDiscount = discounts.Count != 0 ? discounts.Average() : 0.0;
-        Console.WriteLine($"Средняя скидка: {avgDiscount:0.00}%\n");
-    }
+        if (allExhibitionNames.Any(name => name == exhibitionName))
+        {
+            var visitorsQuery =
+                from ticket in _context.Tickets
+                join visitor in _context.Visitors on ticket.VisitorId equals visitor.Id
+                join exhibition in _context.Exhibitions on ticket.ExhibitionId equals exhibition.Id
+                where exhibition.Name == exhibitionName
+                select visitor;
+        
+            var count = await visitorsQuery.CountAsync();
 
-    private async Task GetAllExhibitionNames()
-    {
-        Console.WriteLine("----------- Список всех выставок---------------");
-        
-        var exhibitionsQuery =
-            from exhibition in _context.Exhibitions
-            select exhibition.Name;
-        
-        var exhibitions = await exhibitionsQuery.Distinct().ToListAsync();
-        
-        foreach(var exhibition in exhibitions)
-            Console.WriteLine(exhibition);
-        
-        Console.WriteLine("-----------------------------------------------");
-    }
-    
-    private async Task GetAllVisitorsNames()
-    {
-        Console.WriteLine("----------- Список всех посетителей---------------");
-        
-        var visitorsQuery =
-            from visitor in _context.Visitors
-            select visitor.FullName;
-        
-        var visitors = await visitorsQuery.Distinct().ToListAsync();
-        
-        foreach(var visitor in visitors)
-            Console.WriteLine(visitor);
-        
-        Console.WriteLine("-------------------------------------------------");
+            if (count == 0)
+            {
+                Console.WriteLine($"Посетителей выставки {exhibitionName} не найдено\n");
+            }
+            else
+            {
+                var avgDiscount = await visitorsQuery
+                    .Distinct()
+                    .Select(v => v.Discount)
+                    .AverageAsync();
+
+                Console.WriteLine($"Средняя скидка: {avgDiscount:0.00}%\n");
+            }
+        }
+        else
+        {
+            Console.WriteLine($"Выставки {exhibitionName} не существует");
+        }
     }
 }
